requirements.txt
# Requirements for update_forecast project
# Tested / recommended minimum versions for Python 3.10
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.18
joblib>=1.2.0
xgboost>=1.7.6
scikit-learn>=1.1.3
matplotlib>=3.5.2
statsmodels>=0.13.5
plotly>=5.15.0
openpyxl>=3.1.0

gspread
google-auth
google-auth-oauthlib
google-auth-httplib2

# Optional / utility (uncomment if you need them)
# seaborn>=0.12.0
# jupyterlab
